fix(acp): bound incoming byte-stream frames
